+++ auth-server/src/main/java/com/example/config/ResourceServerConfiguration.java
@@ -1,0 +1,18 @@
+package com.example.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+
+@Configuration
+@EnableResourceServer
+public class ResourceServerConfiguration extends ResourceServerConfigurerAdapter {
+    @Override
+    public void configure(HttpSecurity http) throws Exception {
+        // @formatter:off
+        http.antMatcher("/me").authorizeRequests().anyRequest().authenticated();
+        // @formatter:on
+    }
+}
