+++ auth-server/src/main/resources/static/index.html
@@ -22,6 +22,9 @@
         <div>
             With Github: <a href="/login/github">click here</a>
         </div>
+        <div>
+            With Google: <a href="/login/google">click here</a>
+        </div>
     </div>
     <div class="container authenticated" style="display: none">
         Logged in as: <span id="user"></span>
